# Dedup hash object on parsing