This PR dedups empty hash objects on parsing to reduce retained memory. It reduces ~6MB (~13%) memory on creating `RBS::Environment` (see the benchmark section for more details).



I tried it on https://github.com/ruby/rbs/pull/1950 last year, but this change significantly increased the execution time on Ruby 3.3. So I postponed it after Ruby 3.4.

This problem has been solved in Ruby 3.4, so I'd like to introduce this change to RBS.
https://bugs.ruby-lang.org/issues/20710
This change still has a negative impact for Ruby 3.3 users, but I think it's an acceptable compromise.



This change is a revert of https://github.com/ruby/rbs/pull/1950/commits/3d4c7c1580482e38b90aa29ecb7c6659f9b0868f, but I applied it manually because this change is too old. 


## Benchmarking

### Memory usage

I used `benchmark/memory_new_rails_env.rb` script. It's the memory usage on creating `RBS::Environment` with Rails library.

This change focuses on the retained memory. It reduces retained memory by 6,416,320 bytes of Hash. It's 13% in total, and 94% of Hash!

before

```
Total allocated: 163440076 bytes (1644997 objects)
Total retained:  48278378 bytes (511865 objects)

allocated memory by class
-----------------------------------
  92002896  Hash
  17913828  String
  14623464  Array

retained memory by class
-----------------------------------
  13650690  String
  12324184  RBS::Location
   6796304  Hash
   3736032  Array
```

after

```
Total allocated: 157010796 bytes (1604798 objects)
Total retained:  41862058 bytes (471763 objects)

allocated memory by class
-----------------------------------
  85571056  Hash
  17916388  String
  14623464  Array


retained memory by class
-----------------------------------
  13650690  String
  12324184  RBS::Location
   3736032  Array
   1629040  RBS::Types::Function
   1506080  RBS::MethodType
   1430000  RBS::AST::Members::MethodDefinition
   1097760  RBS::Types::Bases::Any
    833960  RBS::TypeName
    774800  RBS::Types::Function::Param
    753040  RBS::AST::Members::MethodDefinition::Overload
    580480  RBS::Types::ClassInstance
    536240  RBS::Namespace
    413160  RBS::AST::Comment
    379984  Hash
```

### Execution time


I used `benchmark/benchmark_new_env.rb` script. It's time to create `RBS::Environment`.

The result shows that the execution time is mostly not changed. I tried this script several times, and sometimes `before` is faster and vice versa.





before

```
Warming up --------------------------------------
             new_env     1.000 i/100ms
       new_rails_env     1.000 i/100ms
Calculating -------------------------------------
             new_env      8.308 (±12.0%) i/s  (120.37 ms/i) -     81.000 in  10.022728s
       new_rails_env      1.526 (± 0.0%) i/s  (655.13 ms/i) -     16.000 in  10.556997s
```

after


```
Warming up --------------------------------------
             new_env     1.000 i/100ms
       new_rails_env     1.000 i/100ms
Calculating -------------------------------------
             new_env      7.904 (±25.3%) i/s  (126.52 ms/i) -     76.000 in  10.092430s
       new_rails_env      1.675 (± 0.0%) i/s  (596.85 ms/i) -     17.000 in  10.276831s
```